+++ Data/Profunctor/Product/Default/Class.hs
@@ -36,12 +36,12 @@ type DefaultFields p a b = GDefCnstr p (Rep a) (Rep b)
 type DefaultFields' p a = DefaultFields p a a
 
 -- | @'DefaultPConstraints' p a@ expands to the minimal combination of
--- @'Profunctor' p@, @'SemiproductProfunctor' p@, @'SumProfunctor' p@ needed to implement
+-- @'Profunctor' p@, @'SemiproductProfunctor' p@, @'SemisumProfunctor' p@ needed to implement
 -- the instance @'Default' p a a@ for a 'Generic' datatype @a@.
 --
 -- > DefaultPConstraints p Foo =
 -- >   ( SemiproductProfunctor p      -- because Foo has a constructor Bar with many fields
--- >   , SumProfunctor p              -- because Foo has multiple constructors
+-- >   , SemisumProfunctor p              -- because Foo has multiple constructors
 -- >   )
 --
 -- > DefaultConstraints p (a, b) =
@@ -90,7 +90,7 @@ instance (SemiproductProfunctor p, GDefault p f f', GDefault p g g') => GDefault
   type GDefCnstr p (f :*: g) (f' :*: g') = (GDefCnstr p f f', GDefCnstr p g g')
   gdef1 = dimap (\(x :*: y) -> (x, y)) (uncurry (:*:)) $ gdef1 ***! gdef1
 
-instance (SumProfunctor p, GDefault p f f', GDefault p g g') => GDefault p (f :+: g) (f' :+: g') where
+instance (SemisumProfunctor p, GDefault p f f', GDefault p g g') => GDefault p (f :+: g) (f' :+: g') where
   type GDefCnstr p (f :+: g) (f' :+: g') = (GDefCnstr p f f', GDefCnstr p g g')
   gdef1 = dimap sumToEither eitherToSum $ gdef1 +++! gdef1
     where
@@ -106,7 +106,7 @@ type instance GDefPCnstr p U1 = SemiproductProfunctor p
 type instance GDefPCnstr p (M1 i c f) = GDefPCnstr p f
 type instance GDefPCnstr p (K1 i c) = Profunctor p
 type instance GDefPCnstr p (f :*: g) = SemiproductProfunctor p
-type instance GDefPCnstr p (f :+: g) = (SumProfunctor p, GDefPCnstr p f, GDefPCnstr p g)
+type instance GDefPCnstr p (f :+: g) = (SemisumProfunctor p, GDefPCnstr p f, GDefPCnstr p g)
 
 gdef :: (Profunctor p, Generic a, Generic b, GDefault p (Rep a) (Rep b)) => p a b
 gdef = dimap from to gdef1